Fixed: Improve moving file to location where another one exists
Closes #2884

+++ src/NzbDrone.Common/Disk/DiskProviderBase.cs
@@ -278,6 +278,11 @@ namespace NzbDrone.Common.Disk
 
         protected virtual void MoveFileInternal(string source, string destination)
         {
+            if (File.Exists(destination))
+            {
+                throw new FileAlreadyExistsException("File already exists", destination);
+            }
+
             _fileSystem.File.Move(source, destination);
         }
